+++ riscv/src/interrupt/supervisor.rs
@@ -88,11 +88,17 @@ unsafe impl ExceptionNumber for Exception {
     }
 }
 
+/// Checks if a specific core interrupt source is enabled in the current hart (supervisor mode).
+#[inline]
+pub fn is_interrupt_enabled<I: CoreInterruptNumber>(interrupt: I) -> bool {
+    sie::read().is_enabled(interrupt)
+}
+
 /// Disables interrupts for a specific core interrupt source in the current hart (supervisor mode).
 #[inline]
 pub fn disable_interrupt<I: CoreInterruptNumber>(interrupt: I) {
     // SAFETY: it is safe to disable an interrupt source
-    sie::disable_interrupt(interrupt);
+    sie::disable(interrupt);
 }
 
 /// Enables interrupts for a specific core interrupt source in the current hart (supervisor mode).
@@ -107,7 +113,7 @@ pub fn disable_interrupt<I: CoreInterruptNumber>(interrupt: I) {
 /// Ensure that this is called in a safe context where interrupts can be enabled.
 #[inline]
 pub unsafe fn enable_interrupt<I: CoreInterruptNumber>(interrupt: I) {
-    sie::enable_interrupt(interrupt);
+    sie::enable(interrupt);
 }
 
 /// Disables interrupts globally in the current hart (supervisor mode).